Tags are now nested to avoid collisions
