+++ src/main/java/com/paneedah/weaponlib/CustomGui.java
@@ -1,6 +1,7 @@
 package com.paneedah.weaponlib;
 
 import com.paneedah.mwc.instancing.PlayerWeaponInstance;
+import com.paneedah.mwc.instancing.Tags;
 import com.paneedah.weaponlib.animation.AnimationModeProcessor;
 import com.paneedah.weaponlib.animation.gui.AnimationGUI;
 import com.paneedah.weaponlib.config.BalancePackManager;
