+++ src/main/java/com/paneedah/weaponlib/ItemMagazine.java
@@ -2,6 +2,7 @@ package com.paneedah.weaponlib;
 
 import com.paneedah.mwc.instancing.PlayerItemInstanceFactory;
 import com.paneedah.mwc.instancing.PlayerMagazineInstance;
+import com.paneedah.mwc.instancing.Tags;
 import lombok.Getter;
 import net.minecraft.client.model.ModelBase;
 import net.minecraft.entity.EntityLivingBase;
