+++ src/main/java/com/paneedah/weaponlib/MagazineReloadAspect.java
@@ -1,6 +1,7 @@
 package com.paneedah.weaponlib;
 
 import com.paneedah.mwc.instancing.PlayerMagazineInstance;
+import com.paneedah.mwc.instancing.Tags;
 import com.paneedah.mwc.network.NetworkPermitManager;
 import com.paneedah.mwc.utils.MWCUtil;
 import com.paneedah.weaponlib.state.Aspect;
