+++ src/main/java/com/paneedah/weaponlib/WeaponFireAspect.java
@@ -1,6 +1,7 @@
 package com.paneedah.weaponlib;
 
 import com.paneedah.mwc.instancing.PlayerWeaponInstance;
+import com.paneedah.mwc.instancing.Tags;
 import com.paneedah.mwc.network.NetworkPermitManager;
 import com.paneedah.mwc.network.messages.MuzzleFlashMessage;
 import com.paneedah.mwc.network.messages.ShellMessageClient;
