+++ src/main/java/com/paneedah/weaponlib/melee/ItemMelee.java
@@ -2,6 +2,7 @@ package com.paneedah.weaponlib.melee;
 
 import com.paneedah.mwc.instancing.PlayerItemInstance;
 import com.paneedah.mwc.instancing.PlayerItemInstanceFactory;
+import com.paneedah.mwc.instancing.Tags;
 import com.paneedah.weaponlib.*;
 import com.paneedah.weaponlib.ItemAttachment.ApplyHandler2;
 import com.paneedah.weaponlib.crafting.CraftingComplexity;
